requirements.txt
anthropic
tcod
python-dotenv